convert processors and newprojectdialog to ts
